Move formula-related methods to a separate file
